final-pjt/back-socrates/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.7
django-allauth==0.57.0
django-cors-headers==4.3.1
django-seed==0.3.1
djangorestframework==3.14.0
Faker==20.0.3
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
